Move rubocop metrics into their own file
