+++ lib/modularization_statistics/private/datadog_reporter.rb
@@ -5,8 +5,8 @@ require 'dogapi'
 require 'modularization_statistics/private/metrics'
 require 'modularization_statistics/private/metrics/files'
 require 'modularization_statistics/private/metrics/public_usage'
-require 'modularization_statistics/private/metrics/protection_usage'
-require 'modularization_statistics/private/metrics/rubocop_protections_exclusions'
+require 'modularization_statistics/private/metrics/packwerk_checker_usage'
+require 'modularization_statistics/private/metrics/rubocop_usage'
 require 'modularization_statistics/private/metrics/packages'
 require 'modularization_statistics/private/metrics/packages_by_team'
 require 'modularization_statistics/private/metrics/nested_packs'
